SCNG-2 Add layout module. Fix routing.
