+++ src/app/app.module.ts
@@ -1,37 +1,18 @@
 import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
-import {RouterModule, Routes} from '@angular/router';
+import {RouterModule} from '@angular/router';
 import {HttpClient, HttpClientModule} from '@angular/common/http';
 
 import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
 import {TranslateHttpLoader} from '@ngx-translate/http-loader';
 
 import {AppComponent} from './app.component';
-import {LoginModule} from './login/login.module';
-import {LoginComponent} from './login/components/login/login.component';
-import { DashboardModule } from './dashboard/dashboard.module';
-import { DashboardComponent } from './dashboard/components/dashboard/dashboard.component';
+import { LayoutModule } from './layout/layout.module';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 export function HttpLoaderFactory(http: HttpClient) {
   return new TranslateHttpLoader(http);
 }
-
-const appRoutes: Routes = [
-  {
-    path: 'login',
-    component: LoginComponent,
-    pathMatch: 'full'
-  },
-  {
-    path: 'dashboard',
-    component: DashboardComponent,
-    pathMatch: 'full'
-  },
-  {
-    path: '**',
-    redirectTo: '/login'
-  }
-];
 
 @NgModule({
   declarations: [
@@ -39,8 +20,8 @@ const appRoutes: Routes = [
   ],
   imports: [
     BrowserModule,
-    LoginModule,
-    DashboardModule,
+    BrowserAnimationsModule,
+    LayoutModule,
     HttpClientModule,
     TranslateModule.forRoot({
       loader: {
@@ -49,9 +30,7 @@ const appRoutes: Routes = [
         deps: [HttpClient]
       }
     }),
-    RouterModule.forRoot(
-      appRoutes
-    ),
+    RouterModule.forRoot([]),
   ],
   providers: [],
   bootstrap: [AppComponent]
