+++ src/app/dashboard/components/dashboard.component.css
@@ -1,0 +1,11 @@
+.dashboard-welcome-text p {
+    letter-spacing: 0.00938em;
+    font-size: 1rem;
+    line-height: 2;
+    font-weight: 400;
+    color: rgba(0, 0, 0, 0.54);
+}
+
+.dashboard-welcome-text h1 {
+    margin-bottom: -1px;
+}
