--- src/app/dashboard/components/nav-item/nav-item.component.css
@@ -1,4 +1,0 @@
-.icon {
-    opacity: 0.65;
-    margin-right: 25px;
-}
