+++ src/app/dashboard/dashboard.module.ts
@@ -1,36 +1,21 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { DashboardComponent } from './components/dashboard/dashboard.component';
-import { MatToolbarModule } from '@angular/material/toolbar';
-import { MatIconModule } from '@angular/material/icon';
-import { MatSidenavModule } from '@angular/material/sidenav';
-import { MatListModule } from '@angular/material/list';
+import { DashboardComponent } from './components/dashboard.component';
 import { TranslateModule } from '@ngx-translate/core';
 import { RouterModule } from '@angular/router';
-import { AvatarModule} from 'ngx-avatar';
-import { UserInfoComponent } from './components/user-info/user-info.component';
-import { AppbarComponent } from './components/appbar/appbar.component';
-import { NavItemComponent } from './components/nav-item/nav-item.component';
 
 @NgModule({
   declarations: [
-    DashboardComponent,
-    UserInfoComponent,
-    AppbarComponent,
-    NavItemComponent
+    DashboardComponent
   ],
   imports: [
     CommonModule,
-    MatToolbarModule,
-    MatIconModule,
-    MatSidenavModule,
-    MatListModule,
     TranslateModule,
-    RouterModule,
-    AvatarModule
+    RouterModule.forChild([{ path: '', component: DashboardComponent }])
   ],
   exports: [
-    DashboardComponent
+    DashboardComponent,
+    RouterModule
   ]
 })
 export class DashboardModule { }
