+++ src/app/layout/components/layout/layout.component.css
@@ -3,17 +3,23 @@
     width: 270px;
 }
 
-.toolbar {
+.user-info-toolbar {
     height: 100px;
-    background-color: white;
 }
 
-.userinfo {
+.user-info {
     width: 100%;
 }
 
 .feedback-navbar-item {
+    max-width: 270px;
     width: 270px;
     position: fixed;
     bottom: 0;
+}
+
+.content {
+    margin: 24px;
+    padding: 16px;
+    height: calc(100vh - 144px);
 }
