+++ src/app/layout/components/nav-item/nav-item.component.ts
@@ -3,7 +3,7 @@ import { OnInit, Component, Input } from '@angular/core';
 @Component({
     selector: 'app-nav-item',
     template: `
-        <a mat-list-item routerLink={{routerLink}}>
+        <a mat-list-item routerLink={{routerLink}} routerLinkActive="active">
             <mat-icon class="icon">{{iconType}}</mat-icon>
             <ng-content></ng-content>
         </a>
