+++ src/app/layout/components/user-info/user-info.component.ts
